# Faster decimal parsing