@spebern 2023-03-25
I added some failing tests for "+", "-" and "". However, any attempt of fixing this (such as a simple "is_empty" check after handling the sign) has a big performance impact for up to up to 70%, that I cannot explain yet.