@tustvold 2023-03-31
```
123.123                 time:   [42.035 ns 42.094 ns 42.183 ns]
                        change: [-62.228% -61.918% -61.696%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 3 outliers among 100 measurements (3.00%)
  2 (2.00%) high mild
  1 (1.00%) high severe

123.1234                time:   [42.248 ns 42.279 ns 42.312 ns]
                        change: [-61.786% -61.694% -61.616%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high severe

123.1                   time:   [40.929 ns 40.986 ns 41.045 ns]
                        change: [-56.455% -56.286% -56.152%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 2 outliers among 100 measurements (2.00%)
  1 (1.00%) low mild
  1 (1.00%) high mild

123                     time:   [37.682 ns 37.729 ns 37.778 ns]
                        change: [-54.921% -54.837% -54.758%] (p = 0.00 < 0.05)
                        Performance has improved.

-123.123                time:   [42.763 ns 42.796 ns 42.833 ns]
                        change: [-61.283% -61.254% -61.223%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 6 outliers among 100 measurements (6.00%)
  1 (1.00%) low severe
  2 (2.00%) low mild
  3 (3.00%) high mild

-123.1234               time:   [42.858 ns 42.882 ns 42.906 ns]
                        change: [-61.322% -61.283% -61.244%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high severe

-123.1                  time:   [42.256 ns 42.308 ns 42.364 ns]
                        change: [-55.400% -55.347% -55.289%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 2 outliers among 100 measurements (2.00%)
  2 (2.00%) high mild

-123                    time:   [39.039 ns 39.112 ns 39.184 ns]
                        change: [-53.739% -53.670% -53.595%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high mild

0.0000123               time:   [25.752 ns 25.782 ns 25.814 ns]
                        change: [-61.473% -61.428% -61.380%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 2 outliers among 100 measurements (2.00%)
  2 (2.00%) high mild

12.                     time:   [32.498 ns 32.551 ns 32.605 ns]
                        change: [-51.983% -51.800% -51.642%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high severe

-12.                    time:   [33.481 ns 33.546 ns 33.620 ns]
                        change: [-50.743% -50.662% -50.578%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 14 outliers among 100 measurements (14.00%)
  7 (7.00%) low mild
  5 (5.00%) high mild
  2 (2.00%) high severe

00.1                    time:   [23.664 ns 23.690 ns 23.719 ns]
                        change: [-65.650% -65.560% -65.475%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 11 outliers among 100 measurements (11.00%)
  7 (7.00%) high mild
  4 (4.00%) high severe

-00.1                   time:   [24.072 ns 24.082 ns 24.094 ns]
                        change: [-65.453% -65.421% -65.390%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 4 outliers among 100 measurements (4.00%)
  3 (3.00%) high mild
  1 (1.00%) high severe

12345678912345678.1234  time:   [116.16 ns 116.25 ns 116.34 ns]
                        change: [-66.102% -65.917% -65.785%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) low mild

-12345678912345678.1234 time:   [116.72 ns 116.82 ns 116.92 ns]
                        change: [-65.614% -65.559% -65.501%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 4 outliers among 100 measurements (4.00%)
  2 (2.00%) low mild
  2 (2.00%) high mild

99999999999999999.999   time:   [115.77 ns 115.87 ns 115.97 ns]
                        change: [-66.087% -66.017% -65.949%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 8 outliers among 100 measurements (8.00%)
  4 (4.00%) low mild
  4 (4.00%) high mild

-99999999999999999.999  time:   [116.36 ns 116.47 ns 116.59 ns]
                        change: [-65.727% -65.691% -65.654%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high mild

.123                    time:   [25.459 ns 25.484 ns 25.512 ns]
                        change: [-56.754% -56.701% -56.647%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 5 outliers among 100 measurements (5.00%)
  4 (4.00%) high mild
  1 (1.00%) high severe

-.123                   time:   [27.129 ns 27.257 ns 27.384 ns]
                        change: [-54.647% -54.463% -54.300%] (p = 0.00 < 0.05)
                        Performance has improved.

123.                    time:   [38.325 ns 38.371 ns 38.420 ns]
                        change: [-54.232% -54.172% -54.108%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 2 outliers among 100 measurements (2.00%)
  1 (1.00%) low mild
  1 (1.00%) high mild

-123.                   time:   [39.456 ns 39.516 ns 39.576 ns]
                        change: [-53.316% -53.214% -53.105%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 8 outliers among 100 measurements (8.00%)
  4 (4.00%) high mild
  4 (4.00%) high severe
```

Nice work :+1: